joined DB and DI

+++ app/schemas.py
@@ -2,16 +2,18 @@
 
 
 
+from decimal import Decimal
+
 from pydantic import BaseModel, Field, field_validator
 
 
 class OperationRequest(BaseModel):
     wallet_name: str = Field(..., max_length=127)
-    amount: float
+    amount: Decimal
     description: str | None = Field(None, max_length=255)
 
     @field_validator('amount')
-    def amount_must_be_positive(cls, v: float) -> float:
+    def amount_must_be_positive(cls, v: Decimal) -> Decimal:
         # Значение больше нуля
         if v <= 0:
             raise ValueError("Amount must be positive")
@@ -31,7 +33,7 @@ class OperationRequest(BaseModel):
     
 class CreateWalletRequest(BaseModel):
      wallet_name: str = Field(..., max_length=127)
-     initial_balance: float = 0
+     initial_balance: Decimal = 0
      
      @field_validator('wallet_name')
      def name_not_empty(cls, v: str) -> str:
@@ -44,7 +46,7 @@ class CreateWalletRequest(BaseModel):
         # Возвращаем очищенное значение
         return v
      @field_validator('initial_balance')
-     def balance_not_negative(cls, v: float) -> float:
+     def balance_not_negative(cls, v: Decimal) -> Decimal:
          # Значение больше нуля
         if v < 0:
             raise ValueError("Initial balance cantto be negative")
